Speed scales with score
speed = 3 + square-root(score / 10000), so it speeds up by one unit when
your speed hits 10000, 40000, 90000, 160000, . . . . Score multiplier is
increased by 1 when you hit these marks too. Note that because the spawn
rates of power-ups and obstacles are not dependent on speed, they spawn
at a constant rate with respect to time, but spawn less frequently with
respect to distance travelled as speed increases. I think that if they
were to have a constant spawn rate with respect to distance travelled
instead of time the game would become too hard.