Changed blade to shield, cleaned up code, more
The shield image is from the internet, so someone needs to make a new
one.

There were a lot of integers that really should be booleans, so I
changed them. I renamed a lot of variables to be consistent with the
capitalization style of the rest of the code. Variables and methods
should be capitalized like this: rampCoins, magnetTimer; and not like
this: rampcoins, magnettimer. This is for easy readability.

I also made the bracketing of our if statements and methods more
consistent. The opening curly bracket is on the same line as the if
statement or method definition, and is separated from the final
parenthesis by one space.

Snowman, Bear, and SnowMobile now have a "die()" method. You should use
this if you want something to kill them instead of copying and pasting
the code like before.

Shooting the gun is now z, and shooting rockets is now x. This is to
account for potential partial differential homogeneity in the player's
keyboard setup.